refactoring: Moved functions out of TournamentService, into relevant services
