+++ backend/src/main/kotlin/com/turneringsportalen/backend/dto/MatchParticipantDTO.kt
@@ -1,7 +1,7 @@
 package com.turneringsportalen.backend.dto
 
-    data class MatchParticipantDTO(
-        val matchId: Int,
-        val participantId: Int,
-        val index: Int
-    )
+data class MatchParticipantDTO(
+    val matchId: Int,
+    val participantId: Int,
+    val index: Int
+)
