+++ backend/src/main/kotlin/com/turneringsportalen/backend/dto/ParticipantDTO.kt
@@ -8,4 +8,4 @@ data class ParticipantDTO(
     @SerialName("participant_id") val participantId: Int,
     @SerialName("tournament_id") val tournamentId: Int,
     val name: String
-    )
+)
