+++ backend/src/main/kotlin/com/turneringsportalen/backend/services/ParticipantService.kt
@@ -8,9 +8,12 @@ import io.github.jan.supabase.postgrest.from
 @Service
 class ParticipantService(private val client: SupabaseClient) {
 
+    suspend fun addParticipant(participant: Participant) {
+        client.from("participant").insert(participant)
+    }
+
     suspend fun findAllParticipants(): List<Participant>? {
         return client.from("participant").select().decodeList<Participant>()
-
     }
 
     suspend fun findMatchParticipantById(id: Int): Participant? {
@@ -22,15 +25,18 @@ class ParticipantService(private val client: SupabaseClient) {
         }.decodeSingle<Participant>()
     }
 
-    suspend fun addMatchParticipant(participant: Participant) {
-
-        client.from("participant").insert(participant)
+    suspend fun findAllTournamentParticipants(tournamentId: Int): List<Participant>? {
+        return client.from("participant").select {
+            filter {
+                eq("tournament_id", tournamentId)
+            }
+        }.decodeList<Participant>()
     }
 
-    suspend fun deleteMatchParticipant(participant: Participant) {
+    suspend fun deleteParticipant(id: Int) {
         client.from("participant").delete {
             filter {
-                eq("participant_id", participant.participantId ?: 0)
+                eq("participant_id", id)
             }
         }
     }
